Changed plugin removal method & fixed plugin extraction

+++ pulse/package/package_uninstall.py
@@ -51,7 +51,7 @@ def uninstall(package: str, recursive: bool) -> None:
     resource = git_get.get_package_resources(package_path, package_type)
     if resource:
         print(f"Found resource for {re_package[0]}/{re_package[1]} ({package_type})!")
-        remove_resource(resource)
+        remove_resource(package_path, resource, package_type, recursive=recursive)
 
     package_utils.remove_requirements(
         re_package[0],
@@ -106,9 +106,17 @@ def remove_dependencies(dependencies) -> None:
         remove_package(dependence[0], dependence[1])
 
 
-def remove_resource(resource: tuple[str]) -> None:
+def remove_resource(origin_path, resource: tuple[str], package_type, recursive: bool = False) -> None:
     owner, repo, release = resource
-    shutil.rmtree(os.path.join(PLUGINS_PATH, owner, repo))
+    if recursive:
+        shutil.rmtree(os.path.join(PLUGINS_PATH, owner, repo))
+
+    plugin = git_get.get_resource_plugins(origin_path, package_type)
+    plugin_path = os.path.join(
+        REQUIREMENTS_PATH, "plugins", package_utils.get_resource_file("".join(plugin))
+    )
+    if os.path.isfile(plugin_path):
+        os.remove(plugin_path)
 
 
 def remove_package(owner: str, repo: str) -> None:
